+++ src/application_settings/configuring_base.py
@@ -1,47 +1,46 @@
 """Module for handling configuration."""
-from typing import Any, TypeVar
+import sys
+from typing import TypeVar
 
-from pydantic.dataclasses import dataclass
-
-from application_settings.container_base import ContainerBase, FileFormat
+from application_settings.container_base import ContainerBase
 from application_settings.container_section_base import (
     ContainerSectionBase,
     SectionTypeStr,
 )
 
+from .private._file_operations import FileFormat
+
+if sys.version_info >= (3, 12):
+    from typing import override
+else:
+    from typing_extensions import override
+
+
 ConfigT = TypeVar("ConfigT", bound="ConfigBase")
 ConfigT.__doc__ = "Represents ConfigBase and all subclasses"
 
 
-@dataclass(frozen=True)
 class ConfigSectionBase(ContainerSectionBase):
     """Base class for all ConfigSection classes (so that we can bound a TypeVar)"""
 
+    @override
     @classmethod
     def kind_string(cls) -> SectionTypeStr:
-        "Return 'Config'"
+        """Return 'Config'"""
         return "Config"
 
 
-@dataclass(frozen=True)
 class ConfigBase(ContainerBase):
     """Base class for main Config class"""
 
+    @override
     @classmethod
     def kind_string(cls) -> SectionTypeStr:
-        "Return 'Config'"
+        """Return 'Config'"""
         return "Config"
 
+    @override
     @classmethod
     def default_file_format(cls) -> FileFormat:
-        "Return the default file format"
+        """Return the default file format"""
         return FileFormat.TOML
-
-    @classmethod
-    def update(
-        cls: type["ConfigBase"], changes: dict[str, dict[str, Any]]
-    ) -> "ConfigBase":
-        "Update and save the settings with data specified in changes; not meant for config"
-        raise TypeError(
-            "Configs should not be updated runtime; consider converting to settings."
-        )