+++ tests/config_example_module_global.py
@@ -1,0 +1,4 @@
+# pylint: disable=missing-module-docstring
+from .config_example import ExampleInitConfig
+
+test_global: float = ExampleInitConfig.get().field0